+++ lib/sym/app/commands/print_key.rb
@@ -5,7 +5,7 @@ module Sym
     module Commands
       class PrintKey < BaseCommand
         required_options %i(keychain key)
-        incompatible_options %i(examples help version bash_completion)
+        incompatible_options %i(examples help version bash_support)
         try_after :show_examples, :generate_key, :encrypt, :decrypt, :password_protect_key, :keychain_add_key
 
         def execute