+++ spec/integration/sym_cli_spec.rb
@@ -19,12 +19,13 @@ RSpec.describe 'CLI execution', :type => :aruba do
     context 'install bash completion' do
       before &RESET_TEMP_FILE
       after &RESET_TEMP_FILE
-      let(:args) { "--bash-completion #{TEMP_FILE}" }
+      let(:args) { "--bash-support #{TEMP_FILE}" }
 
       it 'should run command' do
         run_simple command
         expect(File.exist?(TEMP_FILE))
-        expect(File.read(TEMP_FILE)).to include(Sym::Constants::Completion::Config[:script])
+        expect(File.read(TEMP_FILE)).to include(Sym::Constants::Bash::Config[:completion][:script])
+        expect(File.read(TEMP_FILE)).to include(Sym::Constants::Bash::Config[:symit][:script])
       end
     end